Destruct Decl Writers at end of translation unit
This allows templates to acquire a definition along the way.

+++ test/test3.hpp
@@ -10,4 +10,4 @@ typedef A::Foo<int> Bar;
 typedef A::Foo<float> Barf;
 }  // namespace B
 
-extern template class A::Foo<double>;
+extern template class A::Foo<float>;
